Add information about Go modules (#430)
* Add information about Go modules

This is a quick pass at including information about Go modules. I kept
things very brief, but this should allow people on Go 1.16 (or later) to
continue with the tutorial.

I mentioned that people should run `go mod init` in all future folders,
but perhaps it would be kinder to add that command to all subsequent
chapters? I couldn't decide, but this is a start.

Thanks again for the book!

* Update hello-world.md

* Fix wording